+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Experimental support for triggering quicksave/load via `ksp::debug::SAVE_LOAD_CONTROL`
   * Current limitations: Triggering a quick-load will terminate all running scripts, including the script triggering it
 * Support `vessel.pilot_input` (capture the flight controls as desired by pilot before mangled by any autopilot)
+* Add UI `container.add_dropdown` element
 
 ## 0.5.7.3 -> 0.5.7.6